refactor(mercury): split aoi.rs by builder family (#244)
aoi.rs was 851 lines (over the 700-line cap). The file held four
distinct AoI packet builder families, each with byte-exact wire
layouts and its own client-side semantics. Split by family so each
file owns one wire group:

  aoi/mod.rs    — BASEMSG_* constants + pack_angle / pack_velocity_xyz
                  shared by builders + module declarations + re-exports
  aoi/create.rs — build_create_entity_base + build_create_entity_cascade
                  (CREATE_ENTITY 0x09 + UPDATE_AVATAR 0x10, then the
                  property cascade with appearance + default NPC stats)
  aoi/leave.rs  — build_entity_invisible (alone) + build_entity_leave
                  (paired with LEAVE_AOI 0x0C)
  aoi/update.rs — build_avatar_update (UPDATE_AVATAR 0x10 relay) +
                  build_forced_position (FORCED_POSITION 0x31 snap)
  aoi/method.rs — build_entity_method_packet (single entity-method call)
  aoi/tests.rs  — existing wire-layout test suite

Public surface preserved: `mercury/mod.rs`'s `pub use aoi::{...}` block
still resolves through the new `aoi/mod.rs` re-exports, so external
callers (`cell::space_manager::queries`, etc.) need no changes.

All 6 aoi tests pass; clippy clean; cargo fmt applied.
